requirements.txt
django >= 6.0
django-environ >= 0.12.0
psycopg2-binary>=2.9.0
pytest>=9.0.2
pytest-django>=4.11.0
pytest-cov>=7.0.0
djangorestframework>=3.16.1
django-cors-headers>=4.0.0
Pillow>=10.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
nltk>=3.8.0
pytesseract>=0.3.10
